requirements.txt
streamlit
langchain_community
pandas
chromadb
langchain_groq
langchain_core
python-docx
python-dotenv
uuid



